mesclar funçao de busca do caio com a minha
